requirements.txt
# FQS - Football Quick Shoot Terminal
# Python Requirements

# Core Framework
textual>=0.47.0
rich>=13.7.0

# HTTP Client
httpx>=0.27.0

# Flask Backend
flask>=3.0.0
flask-cors>=4.0.0

# WebSocket
websockets>=12.0
lomond>=0.3.3

# ---------------------------------------------
# Polymarket API & Trading (from GitHub)
# ---------------------------------------------
# These need to be installed via pip with git URL
git+https://github.com/Polymarket/py-clob-client.git
git+https://github.com/Polymarket/python-order-utils.git

# ---------------------------------------------
# Web3 & Blockchain
# ---------------------------------------------
web3>=6.0.0              # Ethereum interactions & blockchain queries
eth-account>=0.10.0      # Account management & key derivation
eth-utils>=2.3.0         # Ethereum utilities (checksums, encoding, etc.)
eth-typing>=3.5.0        # Type hints for Ethereum
eth-abi>=4.0.0           # ABI encoding/decoding

# ---------------------------------------------
# Data Processing
# ---------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# ---------------------------------------------
# HTTP & API
# ---------------------------------------------
requests>=2.31.0         # Standard HTTP client (sync)
aiohttp>=3.9.0           # Async HTTP client

# ---------------------------------------------
# Utilities
# ---------------------------------------------
python-dotenv>=1.0.0     # Environment variables (.env files)
python-dateutil>=2.8.2   # Date parsing and manipulation
pydantic>=2.5.0          # Data validation and settings
typing-extensions>=4.8.0 # Type hints for Python 3.10+
pyperclip>=1.8.2         # Clipboard access

# ---------------------------------------------
# Cryptography & Auth
# ---------------------------------------------
pycryptodome>=3.19.0     # Crypto primitives for API auth

# ---------------------------------------------
# Logging
# ---------------------------------------------
loguru>=0.7.2            # Advanced logging (better than stdlib)

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
